+++ docs/changes.rst
@@ -17,8 +17,9 @@ Unreleased
 
 **Bugfixes**
 
-- Update crawler ``evilinc``.
+- Update crawler ``evilinc`` to work again.
 - Update crawler ``uvod`` to fetch comment too.
+- Update crawler ``gunshow`` to fetch multiple images per release.
 
 
 1.0.1 (2010-02-23)